sichere user überprüfung im backend

+++ Backend/routes/task/deleteTask.php
@@ -6,13 +6,16 @@ $task = new Task();
 
 if(isset($_SESSION['loggedIn']) && isset($_SESSION['UID'])) {
     if (!empty($TAID)) {
-        try {
-            if($task->deleteTask($TAID)) {
-                echo(json_encode("done"));
-                http_response_code(202);
-            } else http_response_code(422);
-        } catch (PDOException $e) {
-            http_response_code(422);
+        $gotTask = json_decode($task->getTask($TAID));
+        if($gotTask->{'created_by'} === $_SESSION['UID']) {
+            try {
+                if($task->deleteTask($TAID)) {
+                    echo(json_encode("done"));
+                    http_response_code(202);
+                } else http_response_code(422);
+            } catch (PDOException $e) {
+                http_response_code(422);
+            }
         }
     }
 }
